requirements-dev.txt
pyqt5-tools==5.15.4.3.2
pyinstaller==5.7.0
requests==2.28.2
beautifulsoup4==4.11.1
tqdm
sqlacodegen